requirements.txt
asgiref==3.8.1
Django==5.1.6
django-jalali-date==1.1.3
django-render-partial==0.4
jalali_core==1.0.0
jdatetime==5.2.0
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
